requirements.txt
flask
openai
python-dotenv
paramiko
gunicorn